fix(wp-embeds): rename wp-embeds processor
