+++ src/shared/processors/low/index.js
@@ -6,12 +6,12 @@ import gallery from './gallery';
 import wpappbox from './wpappbox';
 import anchor from './anchor';
 import table from './table';
-import removeWpEmbeddedContent from './removeWpEmbeddedContent';
+import removeWpEmbedsInAmp from './removeWpEmbedsInAmp';
 
 export default [
   removeContentAds,
   removeArticles,
-  removeWpEmbeddedContent,
+  removeWpEmbedsInAmp,
   anchor,
   image,
   blockquote,
