+++ src/shared/processors/low/removeWpEmbedsInAmp.js
